Fix the bug #299 and #306. To fix the bug #306, a inline function Gen_Literal_TN_Ex has been added into tn.h. Since the inline function used the marco INT32_MIN and INT32_MAX defined in stdint.h, and those two marcos are enabled only when __STDC_LIMIT_MACROS is defined, we need to add -D__STDC_LIMIT_MACROS in Makefile.gbase(Since many files included tn.h, it's an easier way to define that marco in Makefile). And the source files which has already defined __STDC_LIMIT_MACROS will cause a warnig for macro refiniation. To avoid that warning, those defines are elimiated.
git-svn-id: http://svn.open64.net/svnroot/open64@1014 ead328b1-e903-0410-a19d-eacee4b031f6

+++ trunk/kpro64/be/opt/opt_bdce.cxx
@@ -179,7 +179,7 @@ BITWISE_DCE::Operators_without_dependency(OPERATOR opr)
 // Bitmask_of_size - forms a bit mask of 1's for the number of bits
 // ====================================================================
 UINT64
-BITWISE_DCE::Bitmask_of_size(INT32 vsize)
+BITWISE_DCE::Bitmask_of_size(UINT64 vsize)
 {
   Is_True(vsize != 0, ("BITWISE_DCE::Bitmask_of_size: size cannot be 0"));
   if (vsize >= 64)
